perf(app): style imports improvements
